requirements.txt
click==8.1.2
cmake==3.22.3
dlib==19.23.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.1.1
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysqlclient==2.1.0
numpy==1.22.3
Pillow==9.1.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
WTForms==3.0.1
